Refactoring: first step
First step to refactor irc-slack in a set of packages and commands.
The new structure has one big library `pkg/ircslack` and the command
line tool, `cmd/irc-slack`.
In future commits, the library will be further splitted and cleaned up.
