+++ .gitignore
@@ -1,2 +1,2 @@
 coverage.txt
-irc-slack
+cmd/irc-slack/irc-slack
