+++ go.mod
@@ -17,4 +17,4 @@ require (
 	golang.org/x/sys v0.0.0-20200409092240-59c9f1ba88fa // indirect
 )
 
-go 1.13
+go 1.14
