+++ pkg/ircslack/event_handler.go
@@ -1,4 +1,4 @@
-package main
+package ircslack
 
 import (
 	"fmt"
