+++ pkg/ircslack/irc_channel_test.go
@@ -1,4 +1,4 @@
-package main
+package ircslack
 
 import (
 	"testing"
