+++ pkg/ircslack/irc_context.go
@@ -1,4 +1,4 @@
-package main
+package ircslack
 
 import (
 	"fmt"
