+++ pkg/ircslack/logger.go
@@ -1,0 +1,5 @@
+package ircslack
+
+import "github.com/coredhcp/coredhcp/logger"
+
+var log = logger.GetLogger("ircslack")
