+++ pkg/ircslack/users_test.go
@@ -1,4 +1,4 @@
-package main
+package ircslack
 
 import (
 	"bytes"
