+++ pkg/ircslack/wordwrap.go
@@ -1,4 +1,4 @@
-package main
+package ircslack
 
 import (
 	"strings"
